backend/requirements.txt
# Flask Backend
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-sqlalchemy>=3.1.1
flask-migrate==4.0.5
sqlalchemy>=2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0

# Celery & Redis
celery>=5.3.4
redis==5.0.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.1.0

# AI & Embeddings
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.38.1
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
litellm>=1.0.0

# Utilities
marshmallow==3.20.1
werkzeug==3.0.1
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
openai>=1.0.0
python-pptx>=0.6.21
mammoth>=1.6.0
